conjure-up updates
It was requested that we not use the charmstore as our registry. We have
setup our own registry and doesn't require a `conjure` directory in the
bundle archive any longer. Clean up this bundle by removing unecessary
conjure bits and also document how someone could deploy this bundle via conjure-up.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>